--- lab_1/Task 1 Cyphered text.txt
@@ -1,1 +1,0 @@
-рвиндефлиоееоизиьемсйвциннгатгоискыонвлмйксеоауауклсчсревмоютеомтнетчербяоысоертенбйтрнвынсдгуыиякнкоевеятооомрднднтсйеооеаиотпийможкриинпебелыастопхпоеааепоорркатралдедааенлнипскаласомолпаьтюсодеранедвсапааидаоововацлсибайиаыяохлтглаидвтюеооахесадлрпаутейаебоадемнтйалаойниьичнажиеснддоанндтаиасаиварпрютацаваретьаылнникиюамионмтхоаедиоеоспароссрраиаивтттамсаасреузиыамбуйтовваподннвоааелоееусюатемоитптарепбспихаочрыпотаасеичокараямснллнаанспыьинкадтохзчитаооспуккеамлаоюаиравебртмзчактляитаеаонидмаолатевкёктоаойаонилвариеввхиеаостискчкамтстеаиаапоярхктиаоркабсела
