+++ lab_1/task_1_const.txt
@@ -1,0 +1,1 @@
+63
