+++ lab_1/task_1_original_text.txt
