+++ lab_1/task_1_way.json
@@ -1,0 +1,5 @@
+[
+    "C:\\Proganiy\\OiB\\isb\\lab_1\\task_1_original_text.txt",
+    "C:\\Proganiy\\OiB\\isb\\lab_1\\task_1_key.txt",
+    "C:\\Proganiy\\OiB\\isb\\lab_1\\task_1_const.txt"
+]
