+++ lab_1/task_2_key.txt
