+++ lab_1/task_2_uncyphered_text.txt
@@ -1,0 +1,1 @@
+3А НЕСКОЛЛЬКО ПОСЛЛЕДНИХ ДЕСЯТИЛЛЕТИЙ ТРЕБОВАНИЯ К ИНФОРМАЦИОННОЙ БЕ3ОПАСНОСТИ СУЩЕСТВЕННО И3МЕНИЛЛИСЬ ДО НАЧАЛЛА ШИРОКОГО ИСПОЛЛЬ3ОВАНИЯ АВТОМАТИ3ИРОВАННЫХ СИСТЕМ ОБРАБОТКИ ДАННЫХ БЕ3ОПАСНОСТЬ ИНФОРМАЦИИ ДОСТИГАЛЛАСЬ ИСКЛЛЮЧИТЕЛЛЬНО ФИ3ИЧЕСКИМИ И АДМИНИСТРАТИВНЫМИ МЕРАМИ СПОЯВЛЛЕНИЕМ КОМПЬЮТЕРОВ СТАЛЛА ОЧЕВИДНОЙ НЕОБХОДИМОСТЬ ИСПОЛЛЬ3ОВАНИЯ АВТОМАТИЧЕСКИХ СРЕДСТВ 3АЩИТЫ ФАЙЛЛОВ ДАННЫХ И ПРОГРАММНОЙ СРЕДЫ СЛЛЕДУЮЩИЙ ЭТАП РА3ВИТИЯ АВТОМАТИЧЕСКИХ СРЕДСТВ 3АЩИТЫ СВЯ3АН С ПОЯВЛЛЕНИЕМРАСПРЕДЕЛЛЕННЫХ СИСТЕМ ОБРАБОТКИ ДАННЫХ И КОМПЬЮТЕРНЫХ СЕТЕЙ В КОТОРЫХСРЕДСТВА СЕТЕВОЙ БЕ3ОПАСНОСТИ ИСПОЛЛЬ3УЮТСЯ В ПЕРВУЮ ОЧЕРЕДЬ ДЛЛЯ 3АЩИТЫ ПЕРЕДАВАЕМЫХ ПО СЕТЯМ ДАННЫХ В НАИБОЛЛЕЕ ПОЛЛНОЙ ТРАКТОВКЕ ПОД СРЕДСТВАМИ СЕТЕВОЙ БЕ3ОПАСНОСТИ МЫ БУДЕМ ИМЕТЬ В ВИДУ МЕРЫ ПРЕДОТВРАЩЕНИЯ НАРУШЕНИЙ БЕ3ОПАСНОСТИ КОТОРЫЕ ВО3НИКАЮТ ПРИ ПЕРЕДАЧЕ ИНФОРМАЦИИ ПОСЕТЯМ А ТАКЖЕ МЕРЫ ПО3ВОЛЛЯЮЩИЕ ОПРЕДЕЛЛЯТЬ ЧТО ТАКИЕ НАРУШЕНИЯ БЕ3ОПАСНОСТИ ИМЕЛЛИ МЕСТОЪ
