+++ lab_1/task_2_ways.json
@@ -1,0 +1,5 @@
+[
+    "C:\\Proganiy\\OiB\\isb\\lab_1\\cod9.txt",
+    "C:\\Proganiy\\OiB\\isb\\lab_1\task_2_key.txt",
+    "C:\\Proganiy\\OiB\\isb\\lab_1\\task_2_uncyphered_text.txt"
+]
